chore: Refactor notification
